+++ apps/tup-ui/src/components/auth/index.ts
@@ -1,0 +1,2 @@
+export { default as LoginComponent } from './LoginComponent';
+export { default as LogoutComponent } from './LogoutComponent';
