+++ apps/tup-ui/src/components/profile/index.ts
@@ -1,0 +1,1 @@
+export { default as ProfileComponent } from './ProfileComponent';
